refactor: move articles-list into shared-components/list/
