+++ shared-components/list/main-article-card.tsx
